+++ mlem/core/objects.py
@@ -675,7 +675,8 @@ class DatasetMeta(_WithArtifacts):
         self.dataset = self.reader.read(self.relative_artifacts)
 
     def read_batch(self, batch: int) -> Iterator[DatasetType]:
-        return self.reader.read_batch(self.relative_artifacts, batch)  # type: ignore
+        assert isinstance(self.reader, DatasetReader)
+        return self.reader.read_batch(self.relative_artifacts, batch)
 
     def get_value(self):
         return self.data
