+++ tests/conftest.py
@@ -2,7 +2,7 @@ import os
 import posixpath
 import tempfile
 from pathlib import Path
-from typing import Any, Callable, Iterator, Type
+from typing import Any, Callable, Type
 
 import git
 import pandas as pd
@@ -18,7 +18,6 @@ from mlem import CONFIG
 from mlem.api import init, save
 from mlem.constants import PREDICT_ARG_NAME, PREDICT_METHOD_NAME
 from mlem.contrib.fastapi import FastAPIServer
-from mlem.contrib.pandas import PandasReader, get_pandas_batch_formats
 from mlem.contrib.sklearn import SklearnModel
 from mlem.core.artifacts import LOCAL_STORAGE, FSSpecStorage, LocalArtifact
 from mlem.core.dataset_type import DatasetReader, DatasetType, DatasetWriter
@@ -327,40 +326,6 @@ def dataset_write_read_check(
                 assert custom_eq(new.data, dataset.data)
             else:
                 assert new.data == dataset.data
-
-
-def dataset_write_read_batch_check(
-    dataset: DatasetType,
-    format: str,
-    reader_type: Type[DatasetReader] = None,
-    custom_eq: Callable[[Any, Any], bool] = None,
-):
-    with tempfile.TemporaryDirectory() as tmpdir:
-        BATCH_SIZE = 2
-        storage = LOCAL_STORAGE
-
-        fmt = get_pandas_batch_formats(BATCH_SIZE)[format]
-        art = fmt.write(dataset.data, storage, posixpath.join(tmpdir, "data"))
-        reader = PandasReader(dataset_type=dataset, format=format)
-        artifacts = {"data": art}
-        if reader_type is not None:
-            assert isinstance(reader, reader_type)
-
-        df_iterable: Iterator = reader.read_batch(artifacts, BATCH_SIZE)
-        df = None
-        col_types = None
-        while True:
-            try:
-                chunk = next(df_iterable)
-                if df is None:
-                    df = pd.DataFrame(columns=chunk.columns, dtype=col_types)
-                    col_types = dict(zip(chunk.columns, chunk.dtypes))
-                    df = df.astype(dtype=col_types)
-                df = pd.concat([df, chunk.data], ignore_index=True)
-            except StopIteration:
-                break
-
-        assert custom_eq(df, dataset.data)
 
 
 def check_model_type_common_interface(
